requirements/file_handling.txt
python-magic-bin; platform_system == "Windows"
python-magic; platform_system != "Windows"
PyMuPDF>=1.23
python-docx>=1.1
openpyxl>=3.1
rarfile>=4.1
py7zr>=0.21
pillow>=10.0
opencv-python>=4.9
ffmpeg-python>=0.2
